requirements.txt
adal==1.2.2
applicationinsights==0.11.9
attrs==19.3.0
azure-common==1.1.25
azure-core==1.3.0
azure-graphrbac==0.61.1
azure-identity==1.3.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==8.0.1
azure-mgmt-storage==8.0.0
azureml-automl-core==1.1.5.1
azureml-automl-runtime==1.1.5.2
azureml-core==1.1.5.5
azureml-dataprep==1.3.5
azureml-dataprep-native==14.1.0
azureml-defaults==1.1.5.1
azureml-explain-model==1.1.5
azureml-interpret==1.1.5
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline-core==1.1.5
azureml-telemetry==1.1.5.3
azureml-train-automl-client==1.1.5.1
azureml-train-automl-runtime==1.1.5.3
backports.tempfile==1.0
backports.weakref==1.0.post1
boto3==1.12.31
botocore==1.15.31
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
configparser==3.7.4
contextlib2==0.6.0.post1
convertdate==2.1.3
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
dill==0.3.1.1
distro==1.4.0
docker==4.2.0
docutils==0.15.2
dotnetcore2==2.1.13
ephem==3.7.7.0
fbprophet==0.5
fire==0.3.0
Flask==1.0.3
fusepy==3.0.1
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==19.9.0
holidays==0.10.1
idna==2.9
importlib-metadata==1.6.0
inference-schema==1.0.2
interpret-community==0.6.0
interpret-core==0.1.20
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json-logging-py==0.2
JsonForm==0.0.2
jsonpickle==1.3
jsonschema==3.2.0
JsonSir==0.0.2
keras2onnx==1.6.0
liac-arff==2.4.0
lightgbm==2.3.0
lunardate==0.2.0
MarkupSafe==1.1.1
matplotlib==2.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
msal==1.1.0
msal-extensions==0.1.3
msrest==0.6.11
msrestazure==0.6.3
ndg-httpsclient==0.5.1
nimbusml==1.7.0
numpy==1.16.2
oauthlib==3.1.0
onnx==1.6.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
packaging==20.3
pandas==0.23.4
pathspec==0.7.0
patsy==0.5.1
pmdarima==1.1.1
portalocker==1.6.0
protobuf==3.11.3
psutil==5.7.0
py-cpuinfo==5.0.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypiwin32==223
pyrsistent==0.16.0
pystan==2.19.0.0
python-dateutil==2.8.1
Python-EasyConfig==0.1.7
pytz==2019.3
pywin32==227
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
Resource==0.2.1
rsa==4.0
ruamel.yaml==0.15.89
s3transfer==0.3.3
scikit-learn==0.20.3
scipy==1.1.0
SecretStorage==3.1.2
shap==0.34.0
sip==4.19.13
six==1.14.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
smart-open==1.10.0
statsmodels==0.10.2
termcolor==1.1.0
tornado==6.0.4
tqdm==4.44.1
typing-extensions==3.7.4.1
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==0.16.1
wincertstore==0.2
wrapt==1.11.1
zipp==3.1.0
